move scraper into net and added it to the main namespace
