+++ changelog/5364.breaking.rst
@@ -1,0 +1,1 @@
+``sunpy.util.scraper.Scraper`` has been moved into `sunpy.net`, please update your imports to be ``from sunpy.net import Scraper``.
