+++ sunpy/net/__init__.py
@@ -7,5 +7,6 @@ from sunpy.net import helio as _
 from sunpy.net import jsoc as _
 from sunpy.net import vso as _
 from sunpy.net.fido_factory import Fido
+from sunpy.net.scraper import Scraper
 
-__all__ = ["Fido"]
+__all__ = ["Fido", "Scraper"]
